+++ src/step2-index.js
@@ -1,4 +1,4 @@
-import createGameBox from "./view/web/layers/createGameBox.js";
+import createGameBox from "./view/web/layers/gameBox/createGameBox.js";
 import createHeader from "./view/web/layers/header/header.js";
 import readLottoPriceInput from "./view/web/modules/readLottoPriceInput.js";
