+++ src/view/web/layers/gameBox/createGameBox.js
